net/mlx5e: Fix the minimum required header copy during xmit
The minimum required header was not calculated correctly for TCP
packets. If the user set the min inline mode to L4, TCP_MIN_HLEN (20
bytes) is not enough to pass the HW enforcement check, change the header
size to be calculated by eth_get_headlen function.

Also, add 4 bytes for L2 mode to support packets with vlan.

Issue:745100
Change-Id: I0537d1092b8135978e781a9ab1fceee8970334aa
Fixes: "706be16 net/mlx5e: Check the minimum inline header mode before xmit"
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>