net/mlx5e: Workaround to allow vxlan offload in L2 mode
For some reason OVS adds a match on ECN on flows
with encapsulation.
As a workaround we give ECN the same treatment as
ip frags. We allow offloading flows with ECN == 0
and rely on the hardware to match
those flows even when the IP header is not inlined.
Flows with ECN != 0 are not offloaded.

Issue: None
Change-Id: I99ea025acb86b107ea2458086821c4c537e626ca
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>