ovs: When querying stats, return both SW and HW counters
It takes time for flows to be inserted into the HW and
during this time the packet from this flow go through the
slow path.
When we get a query command for an offloaded flow
we should also take the packets that went through the slow path
into account.

Issue: None
Change-Id: I1edb83f2420bf21d52705574db299852747bf8c4
Fixes: 0b1f74dd ('openvswitch: Add statistics support for HW offloaded [..]')
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>