net/mlx5e: ConnectX-4 firmware support for DCBX
DBCX by default is controlled by firmware. In this
mode, firmware is responsible for reading/sending the TLVs packets
from/to the remote partner. When the driver is loaded, the driver
can leave the DCBX in firmware controlled mode or
switch the DCBX back to host controlled mode.

This patch sets up the infrastructure to support changing
DCBX control mode.

Issue: 741587
Change-Id: Iff8f82c4df2e726a50ca859ce1424f7a8871ad41
Signed-off-by: Huy Nguyen <huyn@mellanox.com>