net/mlx5: Add hold/put rules refcount API
Steering consumers (e.g. sniffer) will need to hold a refcount
on flow rules which weren't created by them until the work on
the rule will be finished.

For that we reveal here an API to hold/put a refcount on a rule and
add a completion mechanism so the rule will not be cleared until
the ref holder will release it.

Issue: 801557
Change-Id: Id849ae60d40af3a46dca4faa239279e841bfd011
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>