net/mlx5: Add hold/put rules refcount API
Steering consumers (e.g. sniffer) will need to hold a refcount
on flow rules which weren't created by them until the work on
the rule will be finished.

For that we reveal here an API to hold/put a refcount on a rule and
add a completion mechanism so the rule will not be cleared until
the ref holder will release it.

Issue: 751790
Change-Id: I2ebe0c2b2a92fb8a1bc9aa358593e3e0761adb5d
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>