net/mlx5e: Clean VLAN VID flow rules on teardown
The driver depends on calls to mlx5e_vlan_rx_kill_vid()
ndo for cleaning VID flow rules.

If such flow rules are not deleted by the adder,
as was found to happen in some cases, the driver fails
to clean flow table resources on teardown.

This patch makes sure that all remaining VID flow rules
are deleted on teardown.

Issue: 801557
Change-Id: Ic9abb391597ecc695cd9eeffef3cd7b50dc7ee09
Fixes: afb736e9330a ('net/mlx5: Ethernet resource handling files')
Signed-off-by: Aviv Heller <avivh@mellanox.com>