net: Update API for VST QinQ support
Intoduce new rtnl UAPI that exposes a list of vlans per VF, giving
the ability for user-space application to specify it for the VF, as an
option to support VST QinQ (802.1ad).
We adjusted IP Link tool to support this option.

For future use cases, the new UAPI supports multipule vlans. For now we
limit the list size to a single vlan in kernel.
Add IFLA_VF_VLAN_LIST in addition to IFLA_VF_VLAN to keep backward
compatibility with older versions of IP Link tool.

Add a vlan protocol parameter to the ndo_set_vf_vlan callback.
We kept 802.1Q as the drivers' default vlan protocol, setting vf vlan
protocol to 802.1ad shall set the vf to VST QinQ mode.

issue: 769060
Change-Id: I65eb93a08e75cb429b2e473d26ceee42d2f632d5
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>