ovs: Use ifindex rather then netdev pointer
Before this patch we stored a pointer to the
offloading netdev in the ovs_flow. When the netdev
was removed the pointer was no longer valid and
using it caused a crash.
To avoid the crash we store an ifindex
and namespace and do a lookup to get the netdev

Issue: 784161
Change-Id: Ibf944cd14ea311254fe3cb243712373877af1f30
Fixes: 0b1f74dd9 ('openvswitch: Add statistics support for HW [...]')
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>