net/mlx5e: Report correct auto negotiation and allow toggling
Previous to this patch auto negotiation was reported off although it was
on by default in hardware. This patch reports the correct information to
ethtool and allows the user to toggle it on/off.

Added another parameter to set port proto function in order to pass
the auto negotiation field to the hardware.

Issue: 795391
Change-Id: Ifa35777a7a748b05af43ed03a7562360dbd0dcd0
Signed-off-by: Gal Pressman <galp@mellanox.com>