net/mlx5e: Add devlink based SRIOV mode changes (legacy --> offloads)
Implement handlers for the devlink commands to get and set the SRIOV
E-Switch mode.

When turning to the offloads mode, we disable the e-switch and enable
it again in the new mode, create the NIC offloads table and create VF reps.

When turning to legacy mode, we remove the VF reps and the offloads
table, and re-initiate the e-switch in it's legacy mode.

The actual creation/removal of the VF reps is done in downstream patches.

Issue: 681039
Change-Id: Ia2e7f5bde16ec546c46946598f193f48cffc9c49
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>