net/mlx5: Introduce nop steering commands
All the nodes in the virtual namespaces exists only in software.
E.g. when we add flow rule with mlx5_add_flow_rule to virtual
namespace, the command of SET_FLOW_TABLE_ENTRY isn't called, the rule
will not be added to the hardware flow tables.

This virtual namespace merely exists in order to describe the RoCE
steering namespace.

Issue: 751790
Change-Id: I6ccbb776d3eac77b9685e214f045bf363e84906d
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>