xfrm: Hold an input state in the flowcache
It turned out that xfrm_state_lock becomes very busy if
many cores processing IPsec in parallel. We can avoid this
if we cache found states in the flowcache. With this we
have to do a slowpath lookup only one time for each flow,
then we have the states in a percpu flowcache. With this,
IPsec scales almost linear with the number of cores (flows).

Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>