net/mlx5: Introduce bulk reading of flow counters
This commit utilize the ability of ConnectX-4 to bulk read flow counters.
Few bulk counter queries could be done instead of issuing thousands of
firmware commands per second to get statistics of all flows set to HW,
such as those programmed when we offload tc filters.

Counters are stored sorted by hardware id, and queried in blocks (id +
number of counters).

Due to hardware requirement, start of block and number of counters in a
block must be four aligned.

Issue: 772961
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
Change-Id: I16627f852663b5f2f20030167e59bf0223a1aada
Signed-off-by: Amir Vadai <amir@vadai.me>