IB/core: Add QP attribute 'flow_entropy'
The QP attribute 'flow_entropy' can be used to help the application
to put QPs into different classes for different purposes.
In RoCE LAG we use this value to put all QPs with the same
flow_entropy on the same port (instead of load balancing it).

This is useful when the application has knowledge about which port
is best for a QP to achive equal balancing or low latency for example.

Issue: 801557
Change-Id: I2d53386273e4ad214a4cf9ca5a4d1266c23caa72
Signed-off-by: Aviv Heller <avivh@mellanox.com>