net/mlx5e: Adjustments in the TC offload code towards reuse for SRIOV
Towards reusing the TC offloads code for an SRIOV use-case, change some of the
helper functions to have _nic in their names so it's clear what's NIC unique
and what's general. Also group together the NIC related helpers so we can easily
branch per the use-case in downstream patch.

Issue: 807806
Change-Id: I62d8f01e0c0210b3dfbaa33364795936e8c0b920
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>