IB/core: Add optional AV attribute for Raw Packet QP in INIT to RTR
In order to configure QoS for Raw Packet QP, the user passes the SL
in the AV attribute when modifying the QP from INIT to RTR state.

This SL is used later on to set the Ethernet Priority of the QP.

Until today, we couldn't pass the AV attribute for Raw Packet QP
because it was blocked in the QP state transition matrix. This
patch adds the AV as optional in INIT to RTR.

issue: 803345
Change-Id: I415d0fc587d186c2e923cfd0442e687dee236e6e
Fixes: 8a51866f0810 ('IB: Add ib_modify_qp_is_ok() library function')
Signed-off-by: Talat Batheesh <talatb@mellanox.com>