net/mlx5e: start/stop all tx queues upon open/close netdev
Start all tx queues (including the inactive ones) when opening the
netdev.
Stop all tx queues (including the inactive ones) when closing the
netdev.
This is a workaround for the false watchdog alarm issue in which the
netdev watchdog is polling all the tx queues which may include inactive
queues and thus once lowering the real tx queues number it will generate
false alarms.

Issue: 807118
Change-Id: Icceb8b909140c1dd1b7bcb2e70ccca84bc3aa240
Fixes: 3947ca185999 ('net/mlx5e: Implement ndo_tx_timeout callback')
Signed-off-by: Mohamad Haj Yahia <mohamad@mellanox.com>