IB/mlx5: Add port counter support for raw packet QP
!!!!!!!!!!!!DON"T PUSH!!!!!!!!!!!!!

Counters weren't updated due to raw packet QPs' traffic since the
counter-id was not associated with the QP. Added support for
associating the q-counter-id with the raw packet QP. The attachment
is done only when changing RQ raw packet QP state from RST to INIT
in modify-RQ command. FW support is required for the above, without
this support raw packet QP counters will not count.

issue: 698171
Change-Id: I78f800d73baa5ddb712a063181ad90f958fa8330
Signed-off-by: Alex Vesker <valex@mellanox.com>