IB/mlx5: Add port counter support for raw packet QP
!!!!!!!!!!!!DON"T PUSH!!!!!!!!!!!!!

Counters weren't updated due to raw packet QPs' traffic since the
counter-id was not associated with the QP. Added support for
associating the q-counter-id with the raw packet QP. The attachment
is done only when changing RQ raw packet QP state from RST to INIT
in modify-RQ command. FW support is required for the above, without
this support raw packet QP counters will not count.

issue: 698171
Change-Id: Ie138d8b6e66dff830e4920d205677247934de563
Signed-off-by: Alex Vesker <valex@mellanox.com>