IB/mlx5: Add port counter support for raw packet QP
Counters weren't updated due to raw packet QPs' traffic since the
counter-id was not associated with the QP. Added support for
associating the q-counter-id with the raw packet QP. The attachment
is done only when changing RQ raw packet QP state from RST to INIT
in modify-RQ command. FW support is required for the above, without
this support raw packet QP counters will not count.

issue: 698171
Change-Id: Id97516c02b794c8c74e61132ba17c28cd3ffbbd5
Signed-off-by: Alex Vesker <valex@mellanox.com>