Add port counter support for raw packet QP
RDMA Counters weren't updated due to raw packet QPs' traffic since
the port counter-id was not associated with the QP. FW support is
required for this feature, without this support raw packet QP
counters will not count. A minor re-factor to the modify raw packet
QP function was needed to pass the  q-counter-id to modify RQ command
function.

issue: 698171
Change-Id: Ic0d976772536c5f248fac5b8c31dbb7653325431
Signed-off-by: Alex Vesker <valex@mellanox.com>