Extend flow steering IPv4 and IPv6 specification.
This patch set extends the IPv4 and IPv6 flow specification.

New fields in IPv4:
	1. Type of Service
	2. Time to Live
	3. Flags
	4. Protocol

New fields in IPv6:
	1. Traffic Class
	2. Flow Label
	3. Next Header
	4. Hop Limit

The first patch in this series add support to extend the flow
steering specification. Previously, user flow specification
and kernel flow specification were must to be equal in their size.

To support extensions, we change the flow specifications conditions
test to be as following:

1. Every mask field that is given by the user and doesn't exist in
   kernel is validated to be zero. Thus, new user-space would only fail
   if it's using (i.e field_mask != 0) a field which doesn't exist in
   the current kernel version.
2. All types are known by the kernel.
3. If the kernel is newer than the user-space, all fields (and by
   that, their masks) weren't given by the user are assumed as zeros
   and ignored.
4. The ABI declares filters as 32bit aligned. This is necessary to
   avoid unaligned access when reading the header.

This change is of course backward compatible.

Issue: 803798
Change-Id: Ie875130451fc52286089401f944382ce33d4f70c
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>