Cover letter: Add packet pacing support for IB verbs
When sending from a 10G host to a 1G host, it is easy to overrun
the receiver, leading to packet loss and traffic backing off.
Similar problems occur when a 10G host sends data to a sub-10G
virtual circuit, or a 40G host sending to a 10G host. Packet pacing
could control packet injection rate and reduces network congestion to
maximize throughput & minimize network latency.

Packet pacing is a rate limiting and shaping for a QP (SQ for RAW QP),
set and change the rate is done by modifying QP. This series of patch
made the following high level changes:

1. Report rate limit capabilities through user data. Reported capabilities
   include: The maximum and minimum rate limit in kbps supported by packet
   pacing; Bitmap showing which QP types are supported by packet pacing
   operation.
2. Extend modify QP interface for growing attributes. Add rate limit
   support to the extended interface.
3. Enable mlx5-based hardware to be able to update the rate limit for
   RAW QP packet.

This patch series depend on the following patches:

Add port counter support for raw packet QP
[1] http://www.spinics.net/lists/linux-rdma/msg39607.html

Issue: 746177
Change-Id: I1c2cb98094804f4fa2c0a50a88511974962d47e8
Signed-off-by: Bodong Wang <bodong@mellanox.com>